+++ src/main/java/com/workaudit/model/Developer.java
@@ -1,15 +1,17 @@
 package com.workaudit.model;
 
+import java.math.BigDecimal;
+
 public class Developer {
     private String name;
-    private double tarifa;
+    private BigDecimal tarifa;
 
-    public Developer(String name, double tarifa) {
+    public Developer(String name, BigDecimal tarifa) {
         this.name = name;
         this.tarifa = tarifa;
     }
 
     public String getName() { return name; }
-    public double getTarifa() { return tarifa; }
-    public void setTarifa(double tarifa) { this.tarifa = tarifa; }
+    public BigDecimal getTarifa() { return tarifa; }
+    public void setTarifa(BigDecimal tarifa) { this.tarifa = tarifa; }
 }
